+++ client/src/tests/CourseView.test.tsx
@@ -12,7 +12,7 @@ import coursesService from '../services/courses';
 import instancesService from '../services/instances';
 import AuthContext from '../context/AuthProvider';
 import mockAssessmentModels from './mock-data/mockAssessmentModels';
-import mockAttainments from './mock-data/mockAttainmentsClient';
+import mockAttainments from './mock-data/mockAttainments';
 import mockCourses from './mock-data/mockCourses';
 import mockInstances from './mock-data/mockInstancesWithStringDates';
 import {
